+++ packages/ai/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.62.0] - 2026-03-23
 
 ### Added